Updates for FunKey-S: 1. Added theme support. 2. Added stylus support. 3. Fixed black screen issue when no BIOS files.

+++ assets/funkeys/launch.sh
@@ -7,18 +7,24 @@ export LD_LIBRARY_PATH=$mydir/libs:$LD_LIBRARY_PATH
 
 sv=`cat /proc/sys/vm/swappiness`
 
+cd $mydir
+
+[ ! -f "system/drastic_bios_arm7.bin" ] && exit
+[ ! -f "system/drastic_bios_arm9.bin" ] && exit
+[ ! -f "system/nds_bios_arm7.bin" ] && exit
+[ ! -f "system/nds_bios_arm9.bin" ] && exit
+[ ! -f "system/nds_firmware.bin" ] && exit
+
+if [ -f "libs/libasound.so.2" ]; then
+    rm -rf libs/libasound.so.2
+fi
+
 # 60 by default
 echo 10 > /proc/sys/vm/swappiness
 
 export SDL_VIDEODRIVER=mmiyoo
 export SDL_AUDIODRIVER=alsa
 export EGL_VIDEODRIVER=mmiyoo
-
-cd $mydir
-
-if [ -f "libs/libasound.so.2" ]; then
-    rm -rf libs/libasound.so.2
-fi
 
 rom=`cat rom.txt`
 ./drastic "$rom"
